+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @speechineer/react
 
+## 0.7.2
+
+### Patch Changes
+
+- Updated dependencies [1aa0afd]
+- Updated dependencies [e8dff7a]
+  - @speechineer/js@0.7.2
+
 ## 0.7.1
 
 ### Patch Changes